:triangular_ruler: GfxScene: filename set to class name
